Avoid mem leak of ptr used in getline()
According `man 3 getline`[1], the ptr used to store string should be init
to NULL before put into getline() funct, otherwise it may seen ptr as an
allocated mem space and save read string into it. However, if the ptr
haven't allocated before, string may be saved into wrong mem space and
cause undefine behavior. Thus this patch init line into NULL ptr.

This patch also free line after the parse is done to avoid mem leak.

It is no need to set buf_size to zero in while loop, because according
to `man 3 getline`, the buf_size will be used by getline to see if it
is needed to reallocate the ptr. If readin line's length is bigger than
buf_size, reallocate ptr to a larger size and save the size into
buf_size. If buf_size is set to zero in while loop, it will cause
getline() to reallocate for every time be called. Which is inefficient.

ref:
[1] - https://man7.org/linux/man-pages/man3/getline.3.html