Fix wrong outputs due to unextended signed bits
This commit fix 2 errs related to unextended signed bits:
1. Immediate value cannot be signed value
2. Incorrect result of SHF

1. Immediate value cannot be signed value

In the funct getsrc, the code is changed from

    else getsrc = `SRC ;

to

    else getsrc = {{(WIDTH-ADDRSIZE){`SRC_SIGNED}}, `SRC};

For the original code, if `SRC is a signed val, say -1, getsrc will be
0000_0fff(4095) instead of ffff_ffff(-1). That's because the bit width
of `SRC is 12 bits, but the bit width of getsrc is 32 bits.
We can use signed bit extend  {{(WIDTH-ADDRSIZE){`SRC_SIGNED}}, `SRC}
to solve this prob.
There is same issue in the task execute when `OPCODE is `LD. This commit
also fix the issue in the same manner.

2. Incorrect result of SHF
SHF will fail when execute following instruct:
        SHF A, #-8 // result shoud be 0000A000(left shift 8 bits)
        A:000000A0
The result will be 00000000 instead of 0000A000
Thats because the previous code will parse -8 as 0x0ff8 = 4088 due to
same reason of issue 1. And the instruct was executed as right shift A by
4088 bits.
This issue is fixed as issue 1 is fixed.